Add ME version to firmware dialog
